+++ src/ts/Ads/Managers/AdUnitTracker.spec.ts
@@ -3,8 +3,7 @@ import { LoadApi, LoadApiMock } from 'Core/Native/__mocks__/LoadApi';
 import { StorageApi, StorageApiMock } from 'Core/Native/__mocks__/StorageApi';
 import { TrackableRefreshManager } from 'Ads/Managers/TrackableRefreshManager';
 import { RefreshManager } from 'Ads/Managers/__mocks__/RefreshManager';
-import { AdUnitTracking } from 'Ads/Utilities/ProgrammaticTrackingService';
-import { ProgrammaticTrackingService, ProgrammaticTrackingServiceMock } from 'Ads/Utilities/__mocks__/ProgrammaticTrackingService';
+import { AdUnitTracking, ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { ListenerApi, ListenerMock } from 'Ads/Native/__mocks__/Listener';
 
 describe('AdUnitTracker', () => {
@@ -12,17 +11,15 @@ describe('AdUnitTracker', () => {
     let loadApi: LoadApiMock;
     let storageApi: StorageApiMock;
     let listenerApi: ListenerMock;
-    let pts: ProgrammaticTrackingServiceMock;
     let trackableRefreshManager: TrackableRefreshManager;
 
     beforeEach(() => {
-        pts = new ProgrammaticTrackingService();
         loadApi = LoadApi();
         storageApi = StorageApi();
         listenerApi = ListenerApi();
         trackableRefreshManager = new TrackableRefreshManager();
         trackableRefreshManager.setRefreshManager(RefreshManager());
-        adUnitTracker = new AdUnitTracker('admob', loadApi, storageApi, listenerApi, trackableRefreshManager, pts);
+        adUnitTracker = new AdUnitTracker('admob', loadApi, storageApi, listenerApi, trackableRefreshManager);
     });
 
     describe('when request load for a placement one time', () => {
@@ -31,8 +28,8 @@ describe('AdUnitTracker', () => {
         });
 
         it('should not send any events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(1);
-            expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(1);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
         });
     });
 
@@ -45,10 +42,10 @@ describe('AdUnitTracker', () => {
             });
 
             it('should send metric events', () => {
-                expect(pts.reportMetricEventWithTags).toBeCalledTimes(times);
-                expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(times);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
                 for (let i = 2; i <= times; i++) {
-                    expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(i, AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
+                    expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(i, AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
                 }
             });
         });
@@ -62,9 +59,9 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(2);
-            expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
-            expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.PossibleDuplicateLoadForPlacement, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(2);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.PossibleDuplicateLoadForPlacement, [undefined]);
         });
     });
 
@@ -76,11 +73,11 @@ describe('AdUnitTracker', () => {
         });
 
         it('should not send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(3);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.AttemptToShowAd, [undefined]);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(3, AdUnitTracking.InitialLoadRequest, [undefined]);
-            expect(pts.reportMetricEventWithTags).not.toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(3);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.AttemptToShowAd, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(3, AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).not.toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
         });
     });
 
@@ -97,10 +94,10 @@ describe('AdUnitTracker', () => {
             });
 
             it('should not send metric events', () => {
-                expect(pts.reportMetricEventWithTags).toBeCalledTimes(2);
-                expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
-                expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.InitialLoadRequest, [undefined]);
-                expect(pts.reportMetricEventWithTags).not.toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(2);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.InitialLoadRequest, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).not.toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
             });
         });
     });
@@ -116,9 +113,9 @@ describe('AdUnitTracker', () => {
             });
 
             it('should send metric events', () => {
-                expect(pts.reportMetricEventWithTags).toBeCalledTimes(2);
-                expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
-                expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(2);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.DuplicateLoadForPlacement, [undefined]);
             });
         });
     });
@@ -134,9 +131,9 @@ describe('AdUnitTracker', () => {
             });
 
             it('should send metric events', () => {
-                expect(pts.reportMetricEventWithTags).toBeCalledTimes(2);
-                expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
-                expect(pts.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.PossibleDuplicateLoadForPlacement, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(2);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.InitialLoadRequest, [undefined]);
+                expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledWith(AdUnitTracking.PossibleDuplicateLoadForPlacement, [undefined]);
             });
         });
     });
@@ -151,9 +148,9 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(2);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.AttemptToShowAd, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(2);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.AttemptToShowAd, [undefined]);
         });
     });
 
@@ -166,10 +163,10 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(3);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.FailedToInvalidate, [undefined]);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(3, AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(3);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.InitialLoadRequest, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(2, AdUnitTracking.FailedToInvalidate, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(3, AdUnitTracking.InitialLoadRequest, [undefined]);
         });
     });
 
@@ -185,8 +182,8 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(1);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MissedImpression, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(1);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MissedImpression, [undefined]);
         });
     });
 
@@ -202,8 +199,8 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(1);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MediationShowCall, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(1);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MediationShowCall, [undefined]);
         });
     });
 
@@ -222,8 +219,8 @@ describe('AdUnitTracker', () => {
         });
 
         it('should send metric events', () => {
-            expect(pts.reportMetricEventWithTags).toBeCalledTimes(1);
-            expect(pts.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MissedImpression, [undefined]);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toBeCalledTimes(1);
+            expect(ProgrammaticTrackingService.reportMetricEventWithTags).toHaveBeenNthCalledWith(1, AdUnitTracking.MissedImpression, [undefined]);
         });
     });
 });