+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -1,7 +1,4 @@
-import { Platform } from 'Core/Constants/Platform';
-import { INativeResponse, RequestManager } from 'Core/Managers/RequestManager';
-import { ClientInfo } from 'Core/Models/ClientInfo';
-import { DeviceInfo } from 'Core/Models/DeviceInfo';
+import { MetricInstance } from 'Ads/Networking/MetricInstance';
 
 export enum ProgrammaticTrackingError {
     TooLargeFile = 'too_large_file', // a file 20mb and over are considered too large
@@ -130,151 +127,47 @@ export enum AdUnitTracking {
     AttemptToShowAd = 'ad_unit_attempt_to_show',
     FailedToInvalidate = 'ad_unit_failed_to_invalidate',
     MissedImpression = 'ad_unit_missed_impression',
-    MediationShowCall = 'ad_unit_mediation_show_call'
+    MediationShowCall = 'ad_unit_mediation_show_call',
+    RealMissedImpression = 'ad_unit_real_missed_impression'
 }
 
-type PTSEvent = AdmobMetric | BannerMetric | CachingMetric | ChinaMetric | VastMetric | MraidMetric | MiscellaneousMetric | LoadMetric | ProgrammaticTrackingError | OMMetric | TimingMetric | AUIMetric | AdUnitTracking;
-
-export interface IProgrammaticTrackingData {
-    metrics: IPTSEvent[];
-}
-
-interface IPTSEvent {
-    name: string;
-    value: number;
-    tags: string[];
-}
+export type PTSEvent = AdmobMetric | BannerMetric | CachingMetric | ChinaMetric | VastMetric | MraidMetric | MiscellaneousMetric | LoadMetric | ProgrammaticTrackingError | OMMetric | TimingMetric | AUIMetric | AdUnitTracking;
 
 export class ProgrammaticTrackingService {
-    // Used for manual verification of PRs merged to ads-sdk-diagnostics that are not yet deployed
-    private stagingBaseUrl: string = 'https://sdk-diagnostics.stg.mz.internal.unity3d.com/';
 
-    private metricPath = 'v1/metrics';
-    private timingPath = 'v1/timing';
+    private static _metricInstance: MetricInstance;
 
-    private _platform: Platform;
-    private _request: RequestManager;
-    private _clientInfo: ClientInfo;
-    private _deviceInfo: DeviceInfo;
-    private _countryIso: string;
-    private _batchedEvents: IPTSEvent[];
-
-    constructor(platform: Platform, request: RequestManager, clientInfo: ClientInfo, deviceInfo: DeviceInfo, country: string) {
-        this._platform = platform;
-        this._request = request;
-        this._clientInfo = clientInfo;
-        this._deviceInfo = deviceInfo;
-        this._countryIso = country;
-        this._batchedEvents = [];
-    }
-
-    protected getBaseUrl(): string {
-        return 'https://sdk-diagnostics.prd.mz.internal.unity3d.com/';
-    }
-
-    private createMetricTags(event: PTSEvent, tags: string[]): string[] {
-        const sdkVersion: string = this._clientInfo.getSdkVersionName();
-        return [this.createAdsSdkTag('mevt', event),
-                this.createAdsSdkTag('sdv', sdkVersion),
-                this.createAdsSdkTag('plt', Platform[this._platform])].concat(tags);
-    }
-
-    private createTimingTags(): string[] {
-        return [
-            this.createAdsSdkTag('sdv', this._clientInfo.getSdkVersionName()),
-            this.createAdsSdkTag('iso', this._countryIso),
-            this.createAdsSdkTag('plt', Platform[this._platform])
-        ];
-    }
-
-    private createErrorTags(event: PTSEvent, adType?: string, seatId?: number): string[] {
-
-        const platform: Platform = this._platform;
-        const osVersion: string = this._deviceInfo.getOsVersion();
-        const sdkVersion: string = this._clientInfo.getSdkVersionName();
-
-        return [
-            this.createAdsSdkTag('eevt', event),
-            this.createAdsSdkTag('plt', Platform[platform]),
-            this.createAdsSdkTag('osv', osVersion),
-            this.createAdsSdkTag('sdv', sdkVersion),
-            this.createAdsSdkTag('adt', `${adType}`),
-            this.createAdsSdkTag('sid', `${seatId}`)
-        ];
-    }
-
-    private createData(event: PTSEvent, value: number, tags: string[]): IProgrammaticTrackingData {
-        return {
-            metrics: [
-                {
-                    name: event,
-                    value: value,
-                    tags: tags
-                }
-            ]
-        };
-    }
-
-    private postToDatadog(metricData: IProgrammaticTrackingData, path: string): Promise<INativeResponse> {
-        const url: string = this.getBaseUrl() + path;
-        const data: string = JSON.stringify(metricData);
-        const headers: [string, string][] = [];
-        headers.push(['Content-Type', 'application/json']);
-        return this._request.post(url, data, headers);
-    }
-
-    public createAdsSdkTag(suffix: string, tagValue: string): string {
-        return `ads_sdk2_${suffix}:${tagValue}`;
-    }
-
-    public reportMetricEvent(event: PTSEvent): Promise<INativeResponse> {
-        return this.reportMetricEventWithTags(event, []);
-    }
-
-    public reportMetricEventWithTags(event: PTSEvent, tags: string[]) {
-        const metricData = this.createData(event, 1, this.createMetricTags(event, tags));
-        return this.postToDatadog(metricData, this.metricPath);
-    }
-
-    public reportErrorEvent(event: PTSEvent, adType: string, seatId?: number): Promise<INativeResponse> {
-        const errorData = this.createData(event, 1, this.createErrorTags(event, adType, seatId));
-        return this.postToDatadog(errorData, this.metricPath);
-    }
-
-    public reportTimingEvent(event: TimingMetric, value: number): Promise<INativeResponse> {
-        // Gate Negative Values
-        if (value > 0) {
-            const timingData = this.createData(event, value, this.createTimingTags());
-            return this.postToDatadog(timingData, this.timingPath);
-        } else {
-            const metricData = this.createData(ProgrammaticTrackingError.TimingValueNegative, 1, this.createMetricTags(event, []));
-            return this.postToDatadog(metricData, this.metricPath);
+    public static initialize(metricInstance: MetricInstance): void {
+        if (!this._metricInstance) {
+            this._metricInstance = metricInstance;
         }
     }
 
-    // TODO: Extend this to all events
-    public batchEvent(metric: TimingMetric, value: number): void {
-        // Curently ignore additional negative time values
-        if (value > 0) {
-            this._batchedEvents = this._batchedEvents.concat(this.createData(metric, value, this.createTimingTags()).metrics);
-        }
-
-        // Failsafe so we aren't storing too many events at once
-        if (this._batchedEvents.length >= 10) {
-            this.sendBatchedEvents();
-        }
+    public static createAdsSdkTag(suffix: string, tagValue: string): string {
+        return this._metricInstance.createAdsSdkTag(suffix, tagValue);
     }
 
-    public async sendBatchedEvents(): Promise<void> {
-        if (this._batchedEvents.length > 0) {
-            const data = {
-                metrics: this._batchedEvents
-            };
-            await this.postToDatadog(data, this.timingPath);
-            this._batchedEvents = [];
-            return;
-        }
-        return Promise.resolve();
+    public static reportMetricEvent(event: PTSEvent): void {
+        this._metricInstance.reportMetricEventWithTags(event, []);
     }
 
+    public static reportMetricEventWithTags(event: PTSEvent, tags: string[]): void {
+        this._metricInstance.reportMetricEventWithTags(event, tags);
+    }
+
+    public static reportErrorEvent(event: PTSEvent, adType: string, seatId?: number): void {
+        this._metricInstance.reportErrorEvent(event, adType, seatId);
+    }
+
+    public static reportTimingEvent(event: TimingMetric, value: number): void {
+        this._metricInstance.reportTimingEvent(event, value);
+    }
+
+    public static batchEvent(metric: TimingMetric, value: number): void {
+        this._metricInstance.batchEvent(metric, value);
+    }
+
+    public static sendBatchedEvents(): void {
+        this._metricInstance.sendBatchedEvents();
+    }
 }