+++ src/ts/Ads/Views/OpenMeasurement/AdmobOpenMeasurementController.ts
@@ -32,9 +32,8 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
     private _deviceInfo: DeviceInfo;
     private _request: RequestManager;
     private _thirdPartyEventManager: ThirdPartyEventManager;
-    private _pts: ProgrammaticTrackingService;
 
-    constructor(platform: Platform, core: ICoreApi, clientInfo: ClientInfo, campaign: AdMobCampaign, placement: Placement, deviceInfo: DeviceInfo, request: RequestManager, omAdViewBuilder: OpenMeasurementAdViewBuilder, thirdPartyEventManager: ThirdPartyEventManager, pts: ProgrammaticTrackingService) {
+    constructor(platform: Platform, core: ICoreApi, clientInfo: ClientInfo, campaign: AdMobCampaign, placement: Placement, deviceInfo: DeviceInfo, request: RequestManager, omAdViewBuilder: OpenMeasurementAdViewBuilder, thirdPartyEventManager: ThirdPartyEventManager) {
         super(placement, omAdViewBuilder);
 
         this._platform = platform;
@@ -44,7 +43,6 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
         this._deviceInfo = deviceInfo;
         this._request = request;
         this._thirdPartyEventManager = thirdPartyEventManager;
-        this._pts = pts;
 
         this._omAdSessionId = JaegerUtilities.uuidv4();
 
@@ -77,14 +75,14 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
     public injectVerificationResources(verificationResources: IVerificationScriptResource[]) {
         const omVendors: string[] = [];
         verificationResources.forEach((resource) => {
-            const om = new OpenMeasurement<AdMobCampaign>(this._platform, this._core, this._clientInfo, this._campaign, this._placement, this._deviceInfo, this._request, resource.vendorKey, this._pts);
+            const om = new OpenMeasurement<AdMobCampaign>(this._platform, this._core, this._clientInfo, this._campaign, this._placement, this._deviceInfo, this._request, resource.vendorKey);
             this._omInstances.push(om);
             this.setupOMInstance(om, resource);
             omVendors.push(resource.vendorKey);
         });
         this._campaign.setOMVendors(omVendors);
         this._thirdPartyEventManager.setTemplateValue(ThirdPartyEventMacro.OM_VENDORS, omVendors.join('|'));
-        this._pts.reportMetricEvent(AdmobMetric.AdmobOMInjected);
+        ProgrammaticTrackingService.reportMetricEvent(AdmobMetric.AdmobOMInjected);
     }
 
     public setupOMInstance(om: OpenMeasurement<AdMobCampaign>, resource: IVerificationScriptResource) {
@@ -162,7 +160,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             impressionObject.adView = adView;
 
             this._omInstances.forEach((om) => {
-                this._pts.reportMetricEvent(AdmobMetric.AdmobOMImpression);
+                ProgrammaticTrackingService.reportMetricEvent(AdmobMetric.AdmobOMImpression);
             });
             super.impression(impressionObject);
 
@@ -174,7 +172,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
             };
 
             this._omInstances.forEach((om) => {
-                this._pts.reportMetricEvent(AdmobMetric.AdmobOMImpression);
+                ProgrammaticTrackingService.reportMetricEvent(AdmobMetric.AdmobOMImpression);
             });
             super.impression(impressionObject);
         });
@@ -182,7 +180,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
 
     public sessionStart(sessionEvent: ISessionEvent) {
         super.sessionStart(sessionEvent);
-        this._pts.reportMetricEvent(AdmobMetric.AdmobOMSessionStart);
+        ProgrammaticTrackingService.reportMetricEvent(AdmobMetric.AdmobOMSessionStart);
     }
 
     /**
@@ -190,7 +188,7 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
      */
     public sessionFinish() {
         super.sessionFinish();
-        this._pts.reportMetricEvent(AdmobMetric.AdmobOMSessionFinish);
+        ProgrammaticTrackingService.reportMetricEvent(AdmobMetric.AdmobOMSessionFinish);
         this._omSessionInterfaceBridge.sendSessionFinish();
     }
 }