+++ src/ts/Performance/AdUnits/PerformanceAdUnitWithAutomatedExperimentParametersFactory.ts
@@ -8,7 +8,7 @@ import { IChina } from 'China/IChina';
 import { AnimatedDownloadButtonEndScreen, EndScreenAnimation } from 'Performance/Views/AnimatedDownloadButtonEndScreen';
 import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
 import { AutomatedExperimentsList, ButtonAnimationsExperiment } from 'Ads/Models/AutomatedExperimentsList';
-import { AUIMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { AUIMetric, ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { PerformanceAdUnitParametersFactory } from 'Performance/AdUnits/PerformanceAdUnitParametersFactory';
 
 export class PerformanceAdUnitWithAutomatedExperimentParametersFactory extends PerformanceAdUnitParametersFactory {
@@ -19,7 +19,7 @@ export class PerformanceAdUnitWithAutomatedExperimentParametersFactory extends P
         super(core, core.Ads, china);
         this._automatedExperimentManager = new AutomatedExperimentManager(core);
         this._automatedExperimentManager.initialize(AutomatedExperimentsList).catch(() => {
-            this._programmaticTrackingService.reportMetricEvent(AUIMetric.AutomatedExperimentManagerInitializationError);
+            ProgrammaticTrackingService.reportMetricEvent(AUIMetric.AutomatedExperimentManagerInitializationError);
         });
         this._automatedExperimentManager.beginExperiment();
     }
@@ -45,7 +45,7 @@ export class PerformanceAdUnitWithAutomatedExperimentParametersFactory extends P
             if ((<string[]>Object.values(EndScreenAnimation)).includes(mabDecision)) {
                 endscreenAnimation = <EndScreenAnimation> mabDecision;
             } else {
-                this._programmaticTrackingService.reportMetricEvent(AUIMetric.InvalidEndscreenAnimation);
+                ProgrammaticTrackingService.reportMetricEvent(AUIMetric.InvalidEndscreenAnimation);
             }
         }